add service func for userEntityService

+++ main/src/main/java/app/service/MedtitationAlbumService.java
@@ -33,6 +33,10 @@ public class MedtitationAlbumService {
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     public UUID createAlbum(UserDetails userDetails,
                                        MeditationAlbumRequest meditationAlbumRequest) {
+        return createNewAlbum(userDetails, meditationAlbumRequest).getId();
+    }
+    public MeditationAlbumEntity createNewAlbum(UserDetails userDetails,
+                                                MeditationAlbumRequest meditationAlbumRequest) {
         UserEntity userEntity = userService.getUserByEmail(userDetails.getUsername());
         checkUserTitleAlbums(userDetails, meditationAlbumRequest.getTitle());
 
@@ -45,7 +49,7 @@ public class MedtitationAlbumService {
         entity.setUser(userEntity);
         entity.setMeditations(userAlbumMeditationEntities);
 
-        return meditationAlbumRepository.save(entity).getId();
+        return meditationAlbumRepository.save(entity);
     }
     public MeditationAlbum getAlbum(UserDetails userDetails,
                                     UUID id) {
@@ -69,6 +73,10 @@ public class MedtitationAlbumService {
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     public void deleteAlbumById(UserDetails userDetails, UUID id) {
         MeditationAlbumEntity entity = checkControl(userDetails, id);
+
+        if (entity.getTitle().equals("Мои медитации")) {
+            throw new IllegalArgumentException("нелья удалить альбом по умолчанию");
+        }
 
         meditationAlbumRepository.delete(entity);
     }
@@ -97,6 +105,17 @@ public class MedtitationAlbumService {
         updatedEntity.setId(entity.getId());
 
         return meditationAlbumMapper.meditationAlbumEntityToMeditationAlbum(meditationAlbumRepository.save(updatedEntity));
+    }
+    public void updateAlbumCheckedMeditations(UserDetails userDetails, UUID id, List<UserMeditationEntity> userMeditationEntities) {
+        MeditationAlbumEntity entity = checkControl(userDetails, id);
+
+        List<UserMeditationEntity> userAlbumMeditationEntities = entity.getMeditations();
+        if (userMeditationEntities != null) {
+            userAlbumMeditationEntities = userMeditationEntities;
+        }
+
+        entity.setMeditations(userAlbumMeditationEntities);
+        meditationAlbumRepository.save(entity);
     }
     private MeditationAlbumEntity checkControl(UserDetails userDetails, UUID id) {
         MeditationAlbumEntity entity = programCommons.getAlbumById(id, meditationAlbumRepository);
